fix no.16

+++ asafov.andrew/P3/matrixAutoInitialization.cpp
@@ -1,7 +1,7 @@
 #include "matrixInitiallization.hpp"
 #include <stdexcept>
 
-void asafov::initializationMatrixAuto(std::istream& in, int* mtx, size_t& n, size_t& m)
+void asafov::initializationMatrixAuto(std::istream& in, int* mtx, unsigned long long& n, unsigned long long& m)
 {
   in >> n;
   if (in.fail() || in.eof()) {
