+++ asafov.andrew/P3/matrixInitiallization.hpp
@@ -4,7 +4,7 @@
 #include <fstream>
 namespace asafov
 {
-  void initializationMatrixAuto(std::istream& in, int* mtx, size_t& n, size_t& m);
+  void initializationMatrixAuto(std::istream& in, int* mtx, unsigned long long& n, unsigned long long& m);
   void initializationMatrixDynamic(std::istream& in, int* mtx, unsigned long long& n, unsigned long long& m);
 }
 #endif
